+++ src/main/java/ru/yandex/money/gradle/plugins/gradleproject/ExtensionConfigurator.java
@@ -1,8 +1,13 @@
 package ru.yandex.money.gradle.plugins.gradleproject;
 
+import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
+import org.apache.commons.lang3.StringUtils;
 import org.gradle.api.Project;
+import org.gradle.plugin.devel.GradlePluginDevelopmentExtension;
 import ru.yandex.money.gradle.plugin.architecturetest.ArchitectureTestExtension;
 import ru.yandex.money.gradle.plugins.gradleproject.git.GitManager;
+import ru.yandex.money.gradle.plugins.javapublishing.JavaArtifactPublishExtension;
+import ru.yandex.money.gradle.plugins.javapublishing.JavaArtifactPublishPlugin;
 import ru.yandex.money.gradle.plugins.library.git.expired.branch.settings.EmailConnectionExtension;
 import ru.yandex.money.gradle.plugins.library.git.expired.branch.settings.GitConnectionExtension;
 import ru.yandex.money.gradle.plugins.release.ReleaseExtension;
@@ -15,6 +20,7 @@ import java.util.Arrays;
  * @author Dmitry Komarov
  * @since 05.12.2018
  */
+@SuppressFBWarnings("HARD_CODE_PASSWORD")
 public class ExtensionConfigurator {
     private static final String GIT_EMAIL = "SvcReleaserBackend@yamoney.ru";
     private static final String GIT_USER = "SvcReleaserBackend";
@@ -29,6 +35,33 @@ public class ExtensionConfigurator {
         configureReleasePlugin(project);
         configureArchitectureTestPlugin(project);
     }
+
+    private static String getStringExtProperty(Project project, String propertyName) {
+        String value = (String)project.getExtensions().getExtraProperties().get(propertyName);
+        if (StringUtils.isBlank(value)) {
+            throw new IllegalArgumentException("property " + propertyName + " is empty");
+        }
+        return value;
+    }
+
+    /**
+     * Сконфигурировать публикацию
+     */
+    static void configurePublishPlugin(Project project) {
+        //Создаем extension сами, для того, чтобы выставить очередность afterEvaluate
+        project.getExtensions().create(JavaArtifactPublishPlugin.Companion.getExtensionName(),
+                JavaArtifactPublishExtension.class);
+        project.getExtensions().getExtraProperties().set("pluginId", "");
+        project.getExtensions().getByType(GradlePluginDevelopmentExtension.class).setAutomatedPublishing(false);
+        JavaArtifactPublishExtension publishExtension = project.getExtensions().getByType(JavaArtifactPublishExtension.class);
+        publishExtension.setNexusUser(System.getenv("NEXUS_USER"));
+        publishExtension.setNexusPassword(System.getenv("NEXUS_PASSWORD"));
+        project.afterEvaluate(p -> {
+            publishExtension.setGroupId("ru.yandex.money.gradle.plugins");
+            publishExtension.setArtifactId(getStringExtProperty(project, "pluginId"));
+        });
+    }
+
 
     private static void configureReleasePlugin(Project project) {
         ReleaseExtension releaseExtension = project.getExtensions().getByType(ReleaseExtension.class);